backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1
langchain-groq==0.0.1
faiss-cpu==1.7.4
pdfplumber==0.9.0
python-dotenv==1.0.0
aiofiles==23.2.1
psycopg[binary]==3.1.12
PyPDF2==3.0.1
sqlalchemy==2.0.21
alembic==1.12.0
asyncpg==0.28.0
databases[postgresql]==0.8.0
pydantic==2.4.2
pydantic-settings==2.0.3


